docs(Alert): updates documentation template
