+++ stories/Link/Link.story.mdx
@@ -11,7 +11,7 @@ import {
 
 <Meta 
   title="Foundation/Link"
-  Component={Link}
+  component={Link}
 />
 
 <Header title="Link">
