+++ stories/List/List.story.mdx
@@ -15,7 +15,7 @@ import {
 
 <Meta 
   title="Foundation/List"
-  Component={List}
+  component={List}
 />
 
 <Header title="List">
